functions/requirements.txt
apify-client==1.6.4
beautifulsoup4==4.12.3
firebase_functions~=0.1.0
langchain==0.2.3
langchain-cohere==0.1.7
langchain-community==0.2.4
langchain-groq==0.1.5
langchain-openai==0.1.8
notion-client==2.2.1
python-dotenv==1.0.1